// code-love/src/app/models/profile.interface.ts
export interface ProfileInterface {
  id?: string;
  uid?: string;
  email?: string;
  password?: string;
  name?: string;
  alias?: string;
  image?: string;
  imageUrl?: string;
  age?: number;
  address?: string;
  phone?: string;
  description?: string;
  interest?: string;
  languaje?: string;
  sex?: string;
}